requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
Jinja2==3.1.6
python-multipart==0.0.20
SQLAlchemy==2.0.43
alembic==1.16.4
python-slugify==8.0.4
asyncpg==0.30.0
python-dotenv==1.1.1
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
PyJWT==2.10.1